refactor: move prepare and complete methods under mod
